server/requirements.txt
alabaster==0.7.12
APScheduler==3.7.0
asgiref==3.3.1
attrs==21.2.0
Babel==2.9.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==3.1.3
django-cors-headers==3.5.0
django-docs==0.3.1
django-filter==2.4.0
django-request-logging==0.7.2
django-simple-history==3.0.0
django-stubs==1.7.0
django-taggit==1.3.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-gis==0.16
djangorestframework-simplejwt==4.6.0
djangorestframework-stubs==1.3.0
djoser==2.1.0
docutils==0.16
Faker==8.1.3
filetype==1.0.7
googlemaps==4.4.2
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy==0.790
mypy-extensions==0.4.3
mysqlclient==2.0.1
oauthlib==3.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.0.1
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.11.1
pytest-django==4.2.0
python-dateutil==2.8.1
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-mock==1.9.2
requests-oauthlib==1.3.0
six==1.15.0
snowballstemmer==2.1.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
Sphinx==4.1.1
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-django==0.5.1
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
zipp==3.4.0
